azure-devops-testing/requirements.txt
requests>=2.23.0
azure-devops>=6.0.0b2